review: address Copilot follow-up feedback on PR #795
Two more fixes:

- buildAreaTimeline now passes the live state through to
  registryDeviceClass, restoring the documented fallback path. Some
  integrations expose device_class only as a state attribute (not
  on the entity registry row), and without this fallback the
  timeline would silently emit raw on/off for those entities while
  the buckets translated them, breaking the "one vocabulary"
  guarantee.
- Make the int-bounds casts explicit in optionalIntArg. The
  comparisons against math.MinInt/MaxInt are technically valid via
  Go's untyped-constant rules, but the intent is clearer with
  explicit int64()/float64() casts. On 64-bit platforms (the only
  ones thane targets) the int64 check is a no-op since int ==
  int64; the float64 check correctly rejects values that would
  overflow on conversion.

New test exercises the state-attribute device_class fallback by
constructing a registry entry without device_class and a state
attribute that carries it, asserting the timeline still translates
through semanticState.

+++ internal/state/awareness/area_activity_tool.go
@@ -129,7 +129,10 @@ func optionalIntArg(args map[string]any, key string) (*int, error) {
 		}
 		return &v, nil
 	case int64:
-		if v < math.MinInt || v > math.MaxInt {
+		// On 64-bit platforms int == int64 so the bounds check is a
+		// no-op; on 32-bit it correctly rejects values that overflow
+		// int. Explicit casts make the comparison's type unambiguous.
+		if v < int64(math.MinInt) || v > int64(math.MaxInt) {
 			return nil, fmt.Errorf("%s out of range", key)
 		}
 		if v < 0 {
@@ -144,7 +147,11 @@ func optionalIntArg(args map[string]any, key string) (*int, error) {
 		if v != math.Trunc(v) {
 			return nil, fmt.Errorf("%s must be a whole number", key)
 		}
-		if v < math.MinInt || v > math.MaxInt {
+		// Explicit float64 casts on the int bounds — float64 has a
+		// 53-bit mantissa so very large ints lose precision, but the
+		// comparison still correctly rejects values that would
+		// overflow int when converted.
+		if v < float64(math.MinInt) || v > float64(math.MaxInt) {
 			return nil, fmt.Errorf("%s out of range", key)
 		}
 		if v < 0 {
